Reorganise graphql module by business purpose
