+++ src/domain/graphql/event/eventResolver.ts
