+++ src/domain/graphql/event/eventSchema.ts
