+++ src/domain/graphql/event/upcomingEventsQueryResolver.ts
@@ -1,5 +1,5 @@
 import linked from "../dataSources/linked";
-import createQueryResolver from "./createQueryResolver";
+import createQueryResolver from "../createQueryResolver";
 
 async function resolver(_, { id }, { language }) {
   const events = linked.getUpcomingEvents(id, language);
